requirements.txt
djangorestframework==3.12.4
Django==3.2.16
pandas
xlsxwriter
django-bootstrap5==22.2